requirements.txt
torch
torchvision
Pillow
python-telegram-bot==20.3
python-dotenv